ingest_api/requirements.txt
fastapi
uvicorn
pydantic
confluent-kafka
psycopg2-binary
python-dotenv